requirements.txt
Pillow
pycocotools
tqdm
cython
scipy
setuptools
tensorboard
